Port aw-watcher-vim to Rust
Missing features:

- user command to manually send a heartbeat
- configuration options, the whole plugin assumes a conncetion to a
  locally running activity watch server (on port 5666, which is the
default port if the server is started in testing mode)

nvim-oxi is used as a bridge to the Neovim C API, which allows calling
that API conveniently from Rust. To avoid blocking the Neovim render
thread during network requests a seperate thread is used to send
hearbeats to the aw server. For communication a mpsc channel is used,
provided by tokio which is also used as async executor running in that
thread. The communication with the aw server is handled by the
aw-client-rust crate from the official aw-server-rust Github repo.

With this basis the plan is to improve the code structure and implement the
missing features (essentially config). Further ideas include deeper
integration with Neovim (eg UI for the watchers status), overall better
logging and to improve already existing crates (mainly aw-client-rust and
aw-model) based on knowledge aquired while working on this.