Extract event handling into separate module
With the new `loop` feature in nvim-oxi it is possible to run "async"
tasks in the Neovim thread. This allows the usage of an independent
error channel for transferring any network errors back from the handler
to the Neovim thread (to log them and set the watcher as disconnected).

The dependency on reqwest is needed, because aw-client-rust does not
pub use their Error type.